feat: final render code

+++ common.h
@@ -20,7 +20,7 @@ inline double degrees_to_radians(double degrees)
 inline double random_double()
 {
 	// Returns a random real in [0,1)
-	return rand() / static_cast<double>(RAND_MAX) + 1.0;
+	return rand() / (RAND_MAX + 1.0);
 }
 
 inline double random_double(double min, double max)
